+++ src/MissionControl.Host.AspnetCore/Content/script.ts
@@ -1,7 +1,6 @@
 interface ICliResponse {
     type: string;
     content: string;
-    terminalId: string;
     statusCode: number;
 }
 
@@ -15,27 +14,73 @@ class Arg {
 }
 
 class HostService {
-    async send(cmd: string, args: Array<Arg>): Promise<ICliResponse> {
+    termId:string;
+    constructor(termId:string) {
+        this.termId = termId;
+    }
+
+    async send(cmd: string, args: Array<Arg>, print: (x:string) => void, finish: () => void) {    
         let headerArgs = "";
         for (let item of args) headerArgs += item.key + "=" + item.val + ";";
 
-        const data = await fetch("mc/cmd/" + cmd, {
-            method: "POST",
-            headers: new Headers({
-                "mc.id": "123",
-                "mc.args": headerArgs
+        const fetchResponse = await fetch("mc/cmd/" + cmd, {
+            method : "POST",
+            headers : new Headers({
+                "mc.id" : this.termId,
+                "mc.args" : headerArgs
             })
         });
 
-        const response: ICliResponse = await data.json();
-        return response;
+        const reader = fetchResponse.body.getReader();
+        let response = "";
+        let cursor = 0;
+        
+        // @ts-ignore
+        const stream = new ReadableStream({ start() {
+            function push() {
+                reader.read().then(({done, value}) => {
+                    if (done) {
+                        finish();
+                        return;
+                    }
+                    
+                    let chunk = new TextDecoder("utf-8").decode(value);
+                    //console.log("Received chunk: " + chunk);
+                    response += chunk;
+                    
+                    let begin = response.indexOf("BEGIN>>", cursor);
+                    let end = response.indexOf("<<END", cursor);
+                    
+                    if (begin > -1 && end > -1) {
+                        try {
+                            
+                            let json = response.substring(begin + 7, end);
+                            
+                            //console.log("Trying to parse: " + json);
+                            
+                            let item = <ICliResponse>JSON.parse(json);
+                            if (item.content !== "")
+                                print(item.content + "<br/>");
+
+                            cursor = end + 1;
+                        }
+                        catch (e) {
+                            //console.log("Error parsing json, waiting for the next chunk")
+                        }
+                    }
+
+                    push();
+                });
+            }
+            push();
+        }});
     }
 }
 
 class Parser {
     getArgs(text: string): Array<Arg> {
-        const args = new Array<Arg>();
-        const parts = text.split(" ");
+        const args = [];
+        const parts = text.split(" ");    
         parts.forEach(part => {
             const arg = part.split("=");
             args.push(new Arg(arg[0], arg.length > 1 ? arg[1] : ""));
@@ -48,45 +93,103 @@ class Parser {
 }
 
 class ViewModel {
-    input: HTMLTextAreaElement;
-    view: HTMLDivElement;
+    input : HTMLTextAreaElement;
+    view : HTMLDivElement;
+    
+    history: Array<string> = [];
+    historyCursor: number = -1;
+    
+    parser : Parser;
+    hostService : HostService;
 
-    parser: Parser;
-    hostService: HostService;
-
-    constructor(input: HTMLTextAreaElement, view: HTMLDivElement) {
+    constructor(input: HTMLTextAreaElement, view : HTMLDivElement) {
         this.view = view;
         this.input = input;
         this.parser = new Parser();
-        this.hostService = new HostService();
+        this.hostService = new HostService(Utils.newGuid());
     }
 
-    init(): void {
-        this.input.addEventListener("keypress", (e: KeyboardEvent) => {
+    init():void {
+        this.print(Resources.help);
+        this.input.addEventListener("keypress", (e : KeyboardEvent) => {
             if (e.which === 13) {
                 this.onExecute(e);
                 this.input.value = "";
                 e.preventDefault();
             }
         });
+        this.input.addEventListener("keyup", (e:KeyboardEvent) => this.onKeyUpDown(e));
+        
     }
 
-    private onExecute(e: KeyboardEvent) {
+    private onKeyUpDown(e:KeyboardEvent) {        
+        if ((e.code != 'ArrowUp' && e.code != 'ArrowDown') || this.history.length == 0) {
+            return
+        };
+
+        let isUp = e.code == "ArrowUp";
+        let isDown = !isUp;
+        
+        this.input.value = this.history[this.historyCursor];
+
+        if (isUp && this.historyCursor > 0) {
+            this.historyCursor -= 1;
+        }
+        else if (isDown && this.historyCursor < this.history.length - 1) {
+            this.historyCursor += 1;
+        }
+    }
+    
+    print(text: string) : void {
+        this.view.innerHTML += "<div class='row'><div class='inner'>" + text + "<br/></div></div>";
+    }
+
+    private async onExecute(e: KeyboardEvent) {
         const input = this.input.value;
         const command = this.parser.getCommand(input);
         const args = this.parser.getArgs(input);
 
-        this.hostService
-            .send(command, args)
-            .then(resp => {
-                console.log(resp);
-                this.view.innerHTML += `<div class='row'><div class='inner output ${resp.type}'>
-                    <p class='cmd'>${input}</p><div class='content'>${resp.content.replace(/\r?\n/g, '<br/>')}</div></div>
-                </div>`
+        this.history.push(input);
+        this.historyCursor = this.history.length - 1;
+        
+        if (command === "help") {
+            this.print(Resources.help);
+            return;
+        }
+
+        if(command === "cls") {
+            this.view.innerHTML = "";
+            return;
+        }
+        
+        this.print(input);
+        let inners = document.getElementsByClassName("inner");
+        let last = inners[inners.length - 1];
+        
+        this.input.disabled = true;
+        await this.hostService.send(command, args,  
+            txt => last.innerHTML += txt.replace(/\r?\n/g, "<br/>"),
+            () => {
+                this.input.disabled = false;
+                this.input.focus();    
             });
     }
 }
 
+class Resources {
+    static readonly help : string = "Some help to get you started:<br>\n" +
+        "<b>list-commands</b> will show a list for discovered commands in your app.<br>\n" +
+        "Add <b>--help</b> argument to see description and available parameters. ";
+}
+
+class Utils {
+    static newGuid() : string {
+        return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+            var r = Math.random()*16|0, v = c == 'x' ? r : (r&0x3|0x8);
+            return v.toString(16);
+        });
+    }
+}
 
 window.onload = () => {
     new ViewModel(
@@ -94,3 +197,4 @@ window.onload = () => {
         <HTMLDivElement>document.getElementById("cli-view"))
         .init();
 };
+