+++ src/MissionControl.Host.Core/Responses/ErrorResponse.cs
@@ -4,10 +4,11 @@ namespace MissionControl.Host.Core.Responses
 {
     public sealed class ErrorResponse : CliResponse
     {
-        public ErrorResponse(string message)
+        public ErrorResponse(string message, HttpStatusCode statusCode = HttpStatusCode.BadRequest)
         {
             Content = message;
-            StatusCode = HttpStatusCode.BadRequest;
+            StatusCode = statusCode;
+            IsDone = true;
         }
 
         public override string Type => "error";