+++ RELEASE-NOTES.md
@@ -6,3 +6,9 @@
 - standard bundled commands: list-commands
 - -help arg interception
 - basic display of response
+
+## 0.9.1 alpha - Dec 2018
+
+- multiple responses streaming
+- commands history (up/down arrow keys)
+- UI tweaks, help, cls, server-info commands